🐛 fix: mudando o int para float na linha 7

+++ fernanda-logicaDeProgramacao/aula-7/listas-exercícios/ex001.py
@@ -4,7 +4,7 @@
 numeros = []  # Lista vazia
 
 for i in range(5):
-    numero = int(input(f"Digite o {i+1}º número: "))
+    numero = float(input(f"Digite o {i+1}º número: "))
     numeros.append(numero)
 
 print("\nSua lista de números completa:")
